Привёл все `maxo/dialogs/**/__init__.py` под оригинальные диалоги
- Добавил `CALLBACK_DATA_KEY` и правильно применил его в `IntentMiddleware`
- Перенёс `new_case_field`, `new_magic_selector`, `Selector` в `widgets/common/case.py`
- Добавил импорт `ManagedTime`
