+++ src/maxo/dialogs/api/internal/middleware.py
@@ -2,4 +2,5 @@ STORAGE_KEY = "aiogd_storage_proxy"
 STACK_KEY = "aiogd_stack"
 CONTEXT_KEY = "aiogd_context"
 PAYLOAD_KEY = "aiogd_original_payload"
+CALLBACK_DATA_KEY = "aiogd_original_callback_data"
 EVENT_SIMULATED = "aiogd_event_simulated"
