+++ src/maxo/dialogs/widgets/common/__init__.py
@@ -1,5 +1,6 @@
 from .action import Actionable
 from .base import BaseWidget
+from .case import Selector, new_case_field, new_magic_selector
 from .managed import ManagedWidget
 from .scroll import (
     BaseScroll,
@@ -20,8 +21,11 @@ __all__ = (
     "OnPageChanged",
     "OnPageChangedVariants",
     "Scroll",
+    "Selector",
     "WhenCondition",
     "Whenable",
+    "new_case_field",
+    "new_magic_selector",
     "sync_scroll",
     "true_condition",
 )
