+++ src/maxo/dialogs/widgets/filter_object.py
