+++ src/maxo/dialogs/widgets/kbd/__init__.py
@@ -39,7 +39,7 @@ from .select import (
 )
 from .state import Back, Cancel, Next, Start, SwitchTo
 from .stub_scroll import StubScroll
-from .time import TimeSelect
+from .time import ManagedTimeSelect, TimeSelect
 
 __all__ = (
     "Back",
@@ -66,6 +66,7 @@ __all__ = (
     "ManagedListGroup",
     "ManagedMultiselect",
     "ManagedRadio",
+    "ManagedTimeSelect",
     "ManagedToggle",
     "Multiselect",
     "Next",
